feat: Add reference image support for virtual try-on
Introduces the ability to use a reference image for virtual try-on features. This includes updating the editor state to store a reference image, modifying the Gemini service to accept and process a reference image, and adjusting UI components to accommodate this new functionality. The 'clothes' preset has been removed to make way for this more general-purpose feature.